requirements.txt
transformers
pytorch
pandas
numpy
wandb
huggingface_cli
datasets
accelerate
deepspeed